refactor to abstract ensemble building and add tests
