+++ .gitignore
@@ -133,3 +133,6 @@ dmypy.json
 
 # VS Code
 .vscode/
+
+# Ignore TDC data
+notebooks/data/
